Fix build

+++ src/OrchardCore/OrchardCore.ContentManagement/DefaultContentManagerSession.cs
@@ -4,7 +4,7 @@ namespace OrchardCore.ContentManagement
 {
     public class DefaultContentManagerSession : IContentManagerSession
     {
-        private readonly Dictionary<int, ContentItem> _itemByVersionId = new Dictionary<int, ContentItem>();
+        private readonly Dictionary<long, ContentItem> _itemByVersionId = new Dictionary<int, ContentItem>();
         private readonly Dictionary<string, ContentItem> _publishedItemsById = new Dictionary<string, ContentItem>();
 
         private bool _hasItems;
